Fix BF16 prefix serialization and add fixed-P8 audit runner with shared-chunk coverage prototype
Preserve historical selection defaults and archived results. Add explicit no-coverage ablation, fixed-period anchor-only selector preloading, INT4 input validation, and mean/normalized-mean/max/balanced selection controls over the existing re-prefill runtime.

Validation: 48 new CPU regressions passed; 1000 valid-input legacy selection cases and 100 INT4 byte-equivalence cases passed. No model, GPU, or SSD performance experiments were run. The balanced selector is an experimental CPU reference, not a validated quality or speed improvement.

+++ src/contiguous_fuxian/sparse_qwen_reprefill.py
@@ -114,7 +114,7 @@ def _write_prefix_tensor(path: Path, tensor) -> None:
 
     import torch
 
-    token_major = tensor.detach().permute(0, 2, 1, 3).contiguous().view(torch.uint8).cpu()
+    token_major = tensor.detach().to(torch.bfloat16).permute(0, 2, 1, 3).contiguous().view(torch.uint8).cpu()
     temporary = path.with_suffix(path.suffix + ".tmp")
     with temporary.open("wb") as handle:
         handle.write(token_major.numpy().tobytes())
